requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==4.2.6
django-celery-beat==2.5.0
django-countries==7.5.1
django-filter==23.5
django-modeltranslation==0.18.11
django-rosetta==0.10.0
django-timezone-field==6.1.0
django-translation-flags==1.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
drf-yasg2==1.19.4
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.5
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
polib==1.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-slugify==8.0.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rosetta==0.3
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
